backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
playwright==1.41.0
TikTokApi==6.2.0
youtube-search-python==1.6.6
celery==5.3.6
redis==5.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
instaloader==4.10.3
